+++ thonny/workbench.py
@@ -334,7 +334,7 @@ class Workbench(tk.Tk):
         self.get_menu("edit", _("Edit"))
         self.get_menu("view", _("View"))
         self.get_menu("run", _("Run"))
-        self.get_menu("device", _("Device"))
+        self.get_menu("tempdevice", _("Device"))
         self.get_menu("tools", _("Tools"))
         self.get_menu("help", _("Help"))
 
@@ -772,6 +772,11 @@ class Workbench(tk.Tk):
         Returns:
             None
         """
+
+        # Temporary solution for plug-ins made for versions before 3.2
+        if menu_name == "device":
+            menu_name = "tools"
+            group = 150
 
         # store command to be published later
         self._commands.append(